+++ docx/__init__.py
@@ -2,7 +2,7 @@
 
 from docx.api import Document  # noqa
 
-__version__ = '0.8.4'
+__version__ = '0.8.5'
 
 
 # register custom Part classes with opc package reader