+++ HISTORY.rst
@@ -9,6 +9,7 @@ NEXT
 - Fix #149: KeyError on Document.add_table()
 - Fix #78: feature: add_table() sets cell widths
 - Add #106: feature: Table.direction (i.e. right-to-left)
+- Add #102: feature: add CT_Row.trPr
 
 
 0.8.4 (2015-02-20)